requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
rapidfuzz>=3.0.0
openai>=1.0.0
chardet>=5.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Authentication
bcrypt>=4.0.0
extra-streamlit-components>=0.1.60

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
